requirements.txt
boto3>=1.34.0
botocore>=1.34.0
pandas>=2.0.0
openpyxl>=3.1.0
jinja2>=3.1.0
click>=8.1.0
pyyaml>=6.0
tabulate>=0.9.0
colorama>=0.4.6
tqdm>=4.66.0
pytest>=7.4.0
pytest-mock>=3.11.0
moto>=4.2.0
